backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
httpx
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]